Removed my Placeholder object+scripts from scene
While everything seemed to work in the unity editor, it did not work in the Hololens emulator. Placeholder script would only work once to place the chest. then it would give errors in the debugger. Also with the Placeholder object in the scene, the TapToPlace scripts on any object wouldn't work, and it would even cause the rotating script to not work if TapToPlace was on the same object. However again, removing the Placeholder object from scene fixed all of this - I think this was because of the outdatedness of some of the methods used - like how I think worldanchors were once implemented in the tap to place script...
This means I still have to figure out a way to attach and remove the worldanchors from objects