FIXED NULL REFERENCED OBJECT BUG
In the BoundingBoxRig.cs script from MRTK, I added the line "Destroy(boxInstance);" as when the output holograms were deleted, it would leave behind the app bar, which slowed down the unity emulator and crashed the hololens emulator and hololens when the root hologram was activated a second time to remove the outputs.

I'm not sure why adding this line works instead of "Destroy (appBarInstance);", as it is the app bar that is left behind, but it works to remove the leftover app bar from view, remove the errors, and no longer slows/ crashes the app.