REWORK ROOT HOLOGRAM INSTANTIATION
Now instead of instantiating and destroying roots from the prefabs when "add/remove station #" is said...
We are now just enabling and disabling the root hologram Game Objects.
Overall this is a smoother process, and there are checks to make sure only one of each root is in the scene anyways.
I think in order to get WorldAnchors to work better we need to do it this way.